requirements.txt
torch
transformers
sentence-transformers
faiss-gpu
gradio
PyMuPDF
nltk
bitsandbytes
